Refactor redundant print inspection
