--- plugin/src/main/kotlin/org/jetbrains/research/pynose/plugin/inspections/pytest/disabled/RedundantPrintTestSmellPytestInspection.kt
@@ -1,16 +1,0 @@
-package org.jetbrains.research.pynose.plugin.inspections.pytest.disabled
-
-import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.diagnostic.Logger
-import com.intellij.psi.PsiElementVisitor
-import org.jetbrains.research.pynose.plugin.inspections.AbstractTestSmellInspection
-import org.jetbrains.research.pynose.plugin.inspections.common.disabled.RedundantPrintTestSmellVisitor
-
-class RedundantPrintTestSmellPytestInspection : AbstractTestSmellInspection() {
-    private val LOG = Logger.getInstance(RedundantPrintTestSmellPytestInspection::class.java)
-
-    override fun buildPytestVisitor(holder: ProblemsHolder, session: LocalInspectionToolSession): PsiElementVisitor {
-        return RedundantPrintTestSmellVisitor(holder, session)
-    }
-}
