--- plugin/src/main/kotlin/org/jetbrains/research/pynose/plugin/inspections/unittest/disabled/RedundantPrintTestSmellUnittestInspection.kt
@@ -1,16 +1,0 @@
-package org.jetbrains.research.pynose.plugin.inspections.unittest.disabled
-
-import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.diagnostic.Logger
-import com.intellij.psi.PsiElementVisitor
-import org.jetbrains.research.pynose.plugin.inspections.AbstractTestSmellInspection
-import org.jetbrains.research.pynose.plugin.inspections.common.disabled.RedundantPrintTestSmellVisitor
-
-class RedundantPrintTestSmellUnittestInspection : AbstractTestSmellInspection() {
-    private val LOG = Logger.getInstance(RedundantPrintTestSmellUnittestInspection::class.java)
-
-    override fun buildUnittestVisitor(holder: ProblemsHolder, session: LocalInspectionToolSession): PsiElementVisitor {
-        return RedundantPrintTestSmellVisitor(holder, session)
-    }
-}
