+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -182,18 +182,10 @@
                          groupPath="Python"
                          groupName="PyNose"
                          enabledByDefault="false"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.unittest.disabled.RedundantPrintTestSmellUnittestInspection"
-                         displayName="Redundant print test (unittest)"
-                         shortName="RedundantPrintTestSmellUnittestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="false"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.pytest.disabled.RedundantPrintTestSmellPytestInspection"
-                         displayName="Redundant print test (pytest)"
-                         shortName="RedundantPrintTestSmellPytestInspection"/>
+                         level="WEAK WARNING"
+                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.universal.RedundantPrintTestSmellInspection"
+                         displayName="Redundant print test"
+                         shortName="RedundantPrintTestSmellInspection"/>
         <localInspection language="Python"
                          groupPath="Python"
                          groupName="PyNose"
