--- plugin/src/main/resources/inspectionDescriptions/RedundantPrintTestSmellPytestInspection.html
@@ -1,9 +1,0 @@
-<html>
-<body>
-
-<b>Redundant Print (pytest)</b> occurs when there is a print statement within the test.
-Print statements are considered to be redundant in unit tests as unit tests are usually executed
-as a part of an automated process with little to no human intervention.
-
-</body>
-</html>
