+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/pytest/disabled/SleepyTestTestSmellPytestInspectionTests.kt
@@ -44,7 +44,7 @@ class SleepyTestTestSmellPytestInspectionTests : AbstractTestSmellInspectionTest
                     "        time.sleep(5)"
         )
         val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WEAK_WARNING })
     }
 
     @Test
