+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/unittest/disabled/RedundantPrintTestSmellUnittestInspectionTests.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.components.service
 import io.mockk.every
 import io.mockk.mockkObject
 import org.jetbrains.research.pynose.plugin.inspections.TestRunnerServiceFacade
+import org.jetbrains.research.pynose.plugin.inspections.universal.RedundantPrintTestSmellInspection
 import org.jetbrains.research.pynose.plugin.util.AbstractTestSmellInspectionTestWithSdk
 import org.jetbrains.research.pynose.plugin.util.TestSmellBundle
 import org.junit.Test
@@ -17,7 +18,7 @@ class RedundantPrintTestSmellUnittestInspectionTests : AbstractTestSmellInspecti
         super.setUp()
         mockkObject(myFixture.project.service<TestRunnerServiceFacade>())
         every { myFixture.project.service<TestRunnerServiceFacade>().getConfiguredTestRunner(any()) } returns "Unittests"
-        myFixture.enableInspections(RedundantPrintTestSmellUnittestInspection())
+        myFixture.enableInspections(RedundantPrintTestSmellInspection())
     }
 
     override fun getTestDataPath(): String {
@@ -30,11 +31,11 @@ class RedundantPrintTestSmellUnittestInspectionTests : AbstractTestSmellInspecti
             "test_file.py", "import unittest\n" +
                     "class SomeClass(unittest.TestCase):\n" +
                     "    def test_something(self):\n" +
-                    "        <warning descr=\"${TestSmellBundle.message("inspections.redundant.print.description")}\">print(\"hello!\")</warning>\n" +
+                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.redundant.print.description")}\">print(\"hello!\")</weak_warning>\n" +
                     "        assert 1 == 1\n" +
-                    "        <warning descr=\"${TestSmellBundle.message("inspections.redundant.print.description")}\">print(\"hello again!\")</warning>"
+                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.redundant.print.description")}\">print(\"hello again!\")</weak_warning>"
         )
-        
+
         myFixture.checkHighlighting()
     }
 
@@ -46,9 +47,9 @@ class RedundantPrintTestSmellUnittestInspectionTests : AbstractTestSmellInspecti
                     "    def test_something(self):\n" +
                     "        print(\"hello!\")"
         )
-        
+
         val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WEAK_WARNING })
     }
 
     @Test
