+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/unittest/disabled/SleepyTestTestSmellUnittestInspectionTests.kt
@@ -49,7 +49,7 @@ class SleepyTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionTe
                     "        time.sleep(5)"
         )
         val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WEAK_WARNING })
     }
 
     @Test
