+++ plugin/src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/redundant_print/unittest/test_redundant_print_multiple.py
@@ -4,19 +4,19 @@ import unittest
 class SomeClass(unittest.TestCase):
 
     def test_something(self):
-        <warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</warning>
+        <weak_warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</weak_warning>
 
     def do_something(self):
         print("print")
 
     def test_something_else(self):
-        <warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</warning>
+        <weak_warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</weak_warning>
 
 
 class OtherClass(unittest.TestCase):
 
     def test_something_other(self):
-        <warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</warning>
+        <weak_warning descr="Print statements are considered to be redundant in unit tests, consider removing them">print("print")</weak_warning>
 
 
 class AnotherClass:
